quick fixes on review comments
