+++ ImageFeed/Core/HTTPMethod.swift
@@ -1,0 +1,6 @@
+enum HTTPMethod: String {
+    case get = "GET"
+    case post = "POST"
+    case put = "PUT"
+    case delete = "DELETE"
+}
